requirements.txt
pandas==2.3.1
PyYAML==6.0.2
Requests==2.32.4
